requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
black==21.9b0
certifi==2020.12.5
chardet==4.0.0
click==8.0.3
Django==3.1.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1
django-phonenumber-field==5.0.0
django-twilio==0.13.1.post0
gunicorn==20.0.2
idna==2.10
mypy-extensions==0.4.3
pathspec==0.9.0
phonenumbers==8.12.20
Pillow==8.3.2
platformdirs==2.4.0
PyJWT==1.7.1
pytz==2021.1
regex==2021.10.23
requests==2.25.1
setproctitle==1.2.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tomli==1.2.1
twilio==6.54.0
typing-extensions==3.10.0.2
urllib3==1.26.4
